Experiment joining in and comparing the CSV file contents with the OFX's, as tables.
Note: In July 2022, I made the parsing routine return *part of* the
OFX file contents as a table and joined both the CSV and OFX files to
see if the CSV was adding data that wasn't present in the OFX. It
turns out the transaction type is more explicit in the CVS file but
can be inferred solely from the OFX file, so decided we'd be better
using just the OFX file (better precision in the reported numbers).
The code is still here though. It could be an idea one day to return
only tables from the extract_from_ofx() and create the transactions in
a separate step. Keeping it as is for now.